chore: Add Driver.executeFind(ParsedFindQuery) method.
This lets lensDataLoader ask the driver to execute its ParsedFindQuery.

PostgresDriver implements the new executeFind method by calling the
same QueryBuilder buildQuery function, but now moved/extracted into
a dedicated buildKnexQuery file/method.
