+++ packages/orm/src/drivers/index.ts
@@ -1,4 +1,4 @@
-export * from "./driver";
+export * from "./Driver";
 export * from "./IdAssigner";
 export * from "./InMemoryDriver";
 export * from "./PostgresDriver";
